feat: Support large indices in 32-bit arch

+++ summa-core/src/components/index_writer_holder.rs
@@ -15,7 +15,7 @@ use crate::Error;
 
 /// Hotcache config
 pub struct HotCacheConfig {
-    pub chunk_size: Option<usize>,
+    pub chunk_size: Option<u64>,
 }
 
 /// Wrap `tantivy::SingleSegmentIndexWriter` and allows to recreate it
